Restore ability to define LDFLAGS via defining %ldflags
Now 'FOO' can be added to LDFLAGS both as:
1) %global ldflags %{ldflags} FOO
2) %global build_ldflags %{build_ldflags} FOO

Previously only (2) worked

This is needed to keep %ldflags macro working as before migration to rpm4.
Note that really %build_flags is used further in the build process, %ldflags is something like a mapping of an old macro to an upstream one in RPM 4.